neural-style-tf-deepdreaming/neural-style-tf-deepdreaming/requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
gradio
tensorflow==2.12.0
tensorflow_hub==0.13.0
opencv-python==4.7.0.72
